+++ src/Config/Registrar.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Michalsn\CodeIgniterHtmx\Config;
+
+class Registrar
+{
+    public static function View(): array
+    {
+        return [
+            'decorators' => ['Michalsn\CodeIgniterHtmx\View\ErrorModalDecorator'],
+        ];
+    }
+}